requirement.txt
jupyter
notebook
pandas
xlrd
nltk
gensim
numpy
scikit-learn
matplotlib
tflearn
pyyaml
h5py